BigInteger
Purpose: To learn the boundaries of int and long (accuracy), to become
more familiar with the java libraries.
There once was a wise servant who saved the life of a princess. The
king promised to pay whatever the servant could dream up. Knowing that
the king loved chess, the servant told the king he would like to have
grains of wheat. One grain on the first square of a chess board. Two
grains on the next. Four on the third, and so on.